Moved playerinputcomponent to separate folder
