+++ StortSpelProjekt/Game/src/main.cpp
@@ -1,5 +1,5 @@
 #include "Engine.h"
-#include "PlayerInputComponent.h"
+#include "Components/PlayerInputComponent.h"
 Scene* GetDemoScene(SceneManager* sm);
 
 int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE, LPSTR, int nCmdShow)
